requirements.txt
accelerate
diffusers==0.14.0
transformers
scipy
ftfy
fire
wandb
safetensors
opencv-python
mediapipe
tensorboard
easydict
onnx
scikit-learn
addict
future
lmdb
Pillow
pyyaml
requests
scikit-image
scipy
tb-nightly
tqdm
yapf
lpips
gdown
Cython>=0.29.28
cmake>=3.22.3
clip
kornia
insightface
onnxruntime
segment_anything
mxnet==1.9.1
mxnet-cu112==1.9.1
numpy==1.23.5
# torch==1.12.1+cu113 --extra-index-url https://download.pytorch.org/whl/cu113
# torchvision==0.13.1+cu113